LayoutTests/platform/ios/imported/w3c/web-platform-tests/selection/selection-direction-on-double-click.tentative-expected.txt

hello, world

FAIL direction returns "none" when there is a double click selection(directionless) assert_equals: expected Text node "hello, world" but got null
